🏗️ Architect: Modularized CatalogMixIn
Decomposed the `apts/observations/catalogs.py` God Mixin into a modular `apts/observations/catalogs/` package. Logic is separated into specialized mixins:
- `MessierCatalogMixIn` (`messier.py`): Messier object property management and visibility logic.
- `NgcCatalogMixIn` (`ngc.py`): NGC/IC catalog handling and visibility filtering.
- `PlanetsCatalogMixIn` (`planets.py`): Solar system body logic, including daytime/sun observation visibility gating.
- `StarsCatalogMixIn` (`stars.py`): Star catalog property management.

The public API `CatalogMixIn` is recomposed in `apts/observations/catalogs/__init__.py` using multiple inheritance to maintain backward compatibility for the `Observation` class and other callers. This refactor improves maintainability by following the Single Responsibility Principle and reducing the size of individual components.

Verified via `tests/unit/test_observations.py`, `tests/unit/test_messier.py`, `tests/unit/test_ngc.py`, `tests/unit/test_solar_objects.py`, and `tests/unit/test_stars.py`.
